// server/db/types.d.ts
export enum RoleType {
	ADMIN = "admin",
	MEMBER = "member",
}

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
	? ColumnType<S, I | undefined, U>
	: ColumnType<T, T | undefined, T>

export type Json = JsonValue

export type JsonArray = JsonValue[]

export type JsonObject = {
	[x: string]: JsonValue | undefined
}

export type JsonPrimitive = boolean | number | string | null

export type JsonValue = JsonArray | JsonObject | JsonPrimitive

export type Timestamp = ColumnType<Date, Date | string, Date | string>
// ...
export interface Users {
	created_at: Generated<Timestamp>
	deactivated: Generated<boolean>
	email: string
	email_verified: Generated<boolean>
	id: Generated<number>
	last_active_at: Timestamp | null
	name: string | null
	role: Generated<RoleType>
	updated_at: Timestamp | null
}